+++ cmd/exporter_test.go
@@ -7,6 +7,7 @@ import (
 
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/urfave/cli"
 )
 
@@ -15,18 +16,16 @@ func TestRunWrongLogLevel(t *testing.T) {
 	set.String("log-level", "foo", "")
 	set.String("log-format", "json", "")
 	fmt.Println(Run(cli.NewContext(nil, set, nil)))
-	if err := Run(cli.NewContext(nil, set, nil)); !strings.HasPrefix(err.Error(), "not a valid logrus Level") {
-		t.Fatalf("Unexpected error : %v", err)
-	}
+	err := Run(cli.NewContext(nil, set, nil))
+	assert.Equal(t, strings.HasPrefix(err.Error(), "not a valid logrus Level"), true)
 }
 
 func TestRunWrongLogType(t *testing.T) {
 	set := flag.NewFlagSet("", 0)
 	set.String("log-level", "debug", "")
 	set.String("log-format", "foo", "")
-	if err := Run(cli.NewContext(nil, set, nil)); !strings.HasPrefix(err.Error(), "Invalid log format") {
-		t.Fatalf("Unexpected error : %v", err)
-	}
+	err := Run(cli.NewContext(nil, set, nil))
+	assert.Equal(t, strings.HasPrefix(err.Error(), "Invalid log format"), true)
 }
 
 func TestRunInvalidConfigFile(t *testing.T) {
@@ -34,7 +33,6 @@ func TestRunInvalidConfigFile(t *testing.T) {
 	set.String("log-level", "debug", "")
 	set.String("log-format", "json", "")
 	set.String("config", "path_does_not_exist", "")
-	if err := Run(cli.NewContext(nil, set, nil)); !strings.HasPrefix(err.Error(), "Couldn't open config file :") {
-		t.Fatalf("Unexpected error : %v", err)
-	}
+	err := Run(cli.NewContext(nil, set, nil))
+	assert.Equal(t, strings.HasPrefix(err.Error(), "Couldn't open config file :"), true)
 }